requirements.txt
# CampusSense Python deps. Python 3.11+.
# Install:  python -m venv .venv && source .venv/Scripts/activate
#           pip install -r requirements.txt
lightgbm==4.6.0
scipy>=1.11
pandas>=2.2
pyarrow>=14
duckdb>=1.0
zeep>=4.2
requests>=2.31
tqdm>=4.66
lxml>=5
python-dateutil>=2.8
joblib>=1.4
anthropic>=0.40
supabase>=2.0
python-dotenv>=1.0
